Ignore values out of range when calculating reviews
